Folder containing Flask code for table and metadata previews.
NOTE: metadata preview relies on reading a JSON file of the
preprocessed metadata that is saved on MY DESKTOP. To make the preview
work, there is a sample metadata file saved as “metadata.json” in this
repository. One must change the path in line 79 of “final_draft.py” to
read that json file to see the preview.